[+FEATURE] Fluid (ViewHelpers): Backported ifAuthenticated & ifHasRole security ViewHelpers
you can now use <f:security.ifAuthenticated> and <f:security.ifHasRole role="foo">
in your Fluid templates to check whether an FE user is currently logged in / belongs to the
specified usergroup.
Note: if role is a numeric value the uid of the usergroup is compared, otherwise the title
of the usergroup.

to deal with BE users you can use the respective be.security.* view helpers

Resolves #9143