[BUGFIX] Consolidate usage of DocumentTemplate
The fix in #40519 changed the behavior of the getDocInstance
method in AbstractBackendViewHelper. Before the global SOBE
was created on the fly, holding the instance of DocumentTemplate
in the doc property.

Since extensions may rely on the existence of SOBE we need to
create it on the fly.

Instead of implicitly creating a stdClass object in case SOBE
is not set, explicitly instantiate this class and set the doc
property with the DocumentTemplate instance.

The usage of the global SOBE is deprecated since now.

Additionally clean up the PageInfoViewHelper in this regard.

Change-Id: Id9ee51a0cd182f33f9d33fa1d2f4d1cc8994c79f
Resolves: #41329
Releases: 6.0
Reviewed-on: http://review.typo3.org/15007
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Helmut Hummel
Tested-by: Helmut Hummel