[BUGFIX] Image viewhelper clears $GLOBALS['TSFE'] in backend context
The image viewhelper replaces the $GLOBALS['TSFE'] array to
simulateFrontendEnvironment if it is called in backend context.
After successful get of the image, it restores the TSFE.

But in case, that the requested image has not been found,
it throws an Tx_Fluid_Core_ViewHelper_Exception, which means
that the TSFE get not restored. This causes errors when other
code tries to access TSFE after the image viewhelper has been
executed.

Resolves: #43446
Releases: 4.5, 4.6, 4.7, 6.0

Change-Id: Ic5b1fc73c74ffa7fca6e36c6f42e3f2084efa7c8
Reviewed-on: http://review.typo3.org/16933
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer