[BUGFIX] Incomplete mock reset in StandaloneViewTest
setUp() fills the makeInstance LIFO via addInstance() with dummy
instances. Some of them are used in the tests and fetched from
this stack during the tests, but not all, which leaves some
instances after this test case is finished.
Other tests can stumble upon that and are feeded with
mocks unexpectedly.
The patch resets the stack in a clean way in tearDown() and
reconstitutes saved singleton object instances afterwards.

Change-Id: I7bea19ec22aec28b1098b7ae40ff4657be34af7f
Fixes: #46695
Related: #38163
Releases: 6.1, 6.0
Reviewed-on: https://review.typo3.org/19310
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn