[BUGFIX] Remove second registration of argument 'placeholder'
Feature #45960 introduced a second registration of the placeholder
argument, causing the Extension Manager 'Get Extensions' function
to throw an exception. The same is true for all extensions using the
textfield viewhelper.

This patch removes the second registration in the render function
parameters list to use the same method in all viewhelpers.

Resolves: #48250
Related: #45960
Releases: 6.2
Change-Id: I25c6c1510c7a38a4a1946205189f0b65a3f4ffe1
Reviewed-on: https://review.typo3.org/20824
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters