[BUGFIX] Failing unit tests due to wrong path handling
In some tests of the StandaloneViewTest case the configured expections
aren't met which results in failing tests on Windows systems. The mock
is configured to return values when calling with giving parameter. The
parameter contains the path to a layout root. During runtime the used
paths are converted to unix like paths but the tests are configured to
expect Windows paths (forward slashes vs. backward slashes). This patch
converts paths to unix like paths before the expectations are configured
to not fail on Windows systems.

Releases: master, 6.2
Resolves: #64656
Change-Id: Ie362cc4e2efa41b6cee068e60313c1bfb22340b2
Reviewed-on: http://review.typo3.org/36455
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>