[FEATURE] Remove ViewHelper xmlns-attributes and specified html tag
In order to include ViewHelper namespaces with xmlns:*-attributes
it is currently required to use e.g. a HTML tag to allow code-completion
by an IDE. The tag itself will also be rendered.
To avoid rending the tag, workarounds are necessary.
One is to use sections, but this approach is counter-intuitive and has
rendering overhead due to extra sections being used. Additionally it is
not possible to use this workaround in layouts.

This patch introduces new functionality to remove all xmlns-attributes
which match ViewHelper namespaces.
For HTML output, we introduce a new attribute, which will cause Fluid
to remove the HTML-tag as well.
The attributes is called >data-namespace-typo3-fluid="true"<

Change-Id: I709d17da86cd79b7b3ba5c83b3f2bd6df8e2bd98
Releases: master
Resolves: #66269
Reviewed-on: http://review.typo3.org/38472
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Daniel Schmidt <daniel@netz-basis.com>
Tested-by: Daniel Schmidt <daniel@netz-basis.com>
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>