[FEATURE] Add MediaViewHelper
In order to comfortably render video, audio and all other file types
with a registered Renderer class (RenderingRegistry introduced with
#61800) in FE, we add a new ViewHelper.

The new MediaViewHelper first checks if there is a Renderer present for
the given file. If not, it will as fallback render a image tag.
This way it is a replacement for the ImageViewHelper in most cases.

Releases: master
Resolves: #66366
Change-Id: I4f92b5f1030238aacb544e45a35189f4e3d0732e
Reviewed-on: http://review.typo3.org/38624
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Sascha Egerer <sascha@sascha-egerer.de>
Reviewed-by: Markus Sommer <markussom@posteo.de>
Tested-by: Markus Sommer <markussom@posteo.de>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>