Revert "[TASK] Make \Be\Menus\ActionMenuItemViewHelper compilable"
This reverts commit 9695759844f2c4981339aff77be50340f8c6b1b6.

Reason for the revert is that initialize() of the abstract is no 
longer called, which takes care of "additionalAttributes" and 
"data-" attributes. The introduced getTagBuilder() method is not 
sufficient. So, basically the rule "no simple migration to compilable 
possible if the VH keeps local state" was violated here.

Change-Id: Ie32bf34f3f230b4fa6c4dafa636ab5e18f400536
Releases: master
Resolves: #66898
Reverts: #66751
Reviewed-on: http://review.typo3.org/39499
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>