[BUGFIX] Disable debug for Fluid Ajax Widget
If TYPO3_CONF_VARS[FE][debug] is enabled then the
parseTime is added at the end of each frontend output.

When having AJAX responses, this might be problematic
as the expected output contains more than what is needed.

Explicitly setting debug=0 for the page type 7076 resolves
the problem in Fluid.

Releases: master
Resolves: #71793
Change-Id: I6295286232dd0b1e4257b4e9a3928837ba6e9a1b
Reviewed-on: https://review.typo3.org/45082
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Tested-by: Frank Nägler <frank.naegler@typo3.org>