[FEATURE] Add Fluid interceptor registration
In Fluid Standalone it is possible to register interceptors to manipulate
the template output on different events. The core should offer a way to
register own interceptor classes. This patch introduces a way to register
classes in TYPO3_CONF_VARS[fluid] which are added to the Fluid Parser
Configuration.

Resolves: #76107
Releases: master
Change-Id: Ib46ed602166161c6d625dcd8419954ae68bff0de
Reviewed-on: https://review.typo3.org/48055
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Peter Foerger <pfoerger@mac.com>
Tested-by: Benni Mack <benni@typo3.org>