[BUGFIX] Make ViewHelperBaseTestcase always available
This is a follow-up commit to f648cdc.

To ease the usage of the core with extensions, with f648cdc we moved the
base test classes into a different place to be able to make them available
when typo3/cms is pulled in as dependency, e.g. for testing extensions or
in projects.

The current .gitattributes file causes that this class was not available
when typo3/cms is pulled in as dependency, e.g. for testing extensions or
in projects.

This is a non breaking change, as the classes will be available in any
case after that.

Resolves: #77205
Releases: master
Change-Id: I6f00ac1cb8c48cf2faea8bae666af691d8a1f7c2
Reviewed-on: https://review.typo3.org/49158
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Tested-by: Helmut Hummel <typo3@helhum.io>