[BUGFIX] Do not strip sub-paths from emulated controller name
This patch solves a special case in the integration with
Fluid. Fluid itself allows an emulated controller action
to be used, which contains a sub-path (controller name
supports this as well). However, the RenderingContext
forced use of pathinfo() on the passed path, which caused
sub-paths to be removed. The patch switches from
pathinfo() to substr() to preserve any slashes in such
controller action names.

Only direct usage of Fluid's API is affected and no
current compatibility is broken by the patch.

Change-Id: I7e1c27d55a44b957aadbace571e50afd7614d8a0
Resolves: #79519
Releases: master
Reviewed-on: https://review.typo3.org/51451
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Philipp Gampe <philipp.gampe@typo3.org>
Tested-by: Philipp Gampe <philipp.gampe@typo3.org>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>