[FEATURE] Allow to set a custom id for fluid widgets
With this patch you can set a customWidgetId parameter in the
widget viewhelper. This id is used in the widget identifier
in addition to the nextWidgetNumber.

With this patch two equal widgets in different content elements
don't influence each other.

A good value for the customWidgetId is the {contentObjectData.uid}.

Example:  <f:widget.paginate customWidgetId="{contentObjectData.uid}" ...

Resolves: #47006
Releases: master
Change-Id: Ief1d38457da73a35e652f11d82f441898e0394d3
Reviewed-on: https://review.typo3.org/51518
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>