Revert "[TASK] Improve TranslateViewHelper compiling"
The micro optimization introduced a bug, fixing the bug
would basically lead to the same code that is within
renderStatic() already. There is then no advantage in
having an inline compiled solution in comparison to a
static call to the class anymore. The initial patch and
its follow up are reverted.

This reverts commit e6f470ebd2ddac7f0c864fd7c18c3d12a9009c30.
This reverts commit 852ad1d9ad6f3bac4c977d2466d1875f9e632e01.

Resolves: #86292
Reverts: #82454
Releases: master
Change-Id: I573a4298b0b9aedf83d51416919204bf910ae4bf
Reviewed-on: https://review.typo3.org/58322
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>