[TASK] Streamline phpdoc annotations in EXT:fluid
EXT:fluid uses @api and @internal phpDoc annotations. Going with a cleaner
approach with marking code as just @internal, and everything not annotated
as @internal is public, is the way to go, also the way PSR-5 is heading.

For EXT:fluid: Every PHP class that had nothing marked, is now @internal,
everything that was @api is now implicitly part of TYPO3 Core API.

Other areas will follow so the API for TYPO3 v9 LTS is clearer than ever.

Resolves: #86507
Releases: master
Change-Id: I5deba0ba2b868743e7eda8f0e32cda40ef008046
Reviewed-on: https://review.typo3.org/58525
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>