[BUGFIX] Let CObjectViewHelper render top level objects again
Making the CObjectViewHelper had the unfortunate side effect,
that top level object paths are not working any more.

Adapt the tests to show the mistake and make the rendering
work in this case.

Additionally now an exception is also thrown, if the given
object path does not resolve to a TypoScript object.

Releases: master
Resolves: #86729
Related: #84345
Change-Id: I686119b480c6fae88be7d704b82094e35789c0d7
Reviewed-on: https://review.typo3.org/58686
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>