[!!!][TASK] Deprecate useCacheHash/noCacheHash
Having Site Handling in place, the option "useCacheHash"
rendered useless as it is added by TYPO3 Core automatically
when necessary with SiteHandling.

Any Fluid Arguments haven't been deprecated as this
was part of the the legacy functionality. Now the options
are still available, but trigger a deprecation message
that the option is superfluous and should be removed.

Calling typolink.useCacheHash will now trigger a
deprecation message.

Resolves: #88406
Releases: master
Change-Id: I2243a335188c3466b8f8f59e8d3e417f13bf854d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/60774
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>