[TASK] Update Fluid Standalone to 2.6.6
A new fluid standalone version 2.6.6 reverted the changes
that had regressions in v2.6.5 and v2.6.4.

This patch updates the root composer.json to use latest fluid
standalone, and marks the affected versions as conflicting.

In addition, raising the requirements to 2.6.4 in the first place
was a mistake for this major version, which is why the main requirement
was lowered to the previous version again.

Composer: composer req typo3fluid/fluid:^2.6.1

Resolves: #89542
Releases: master, 9.5, 8.7
Change-Id: I30a29a3cc07a277bd4a72e83a67a6c3260adc5d7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62140
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Benni Mack <benni@typo3.org>