[BUGFIX] Fix broken Fluid templates
In some versions of PHP there is a PCRE bug, which will be triggered when
using an unquoted array syntax in Fluid templates, leading to an error
and possibly not able to display the rendered template.

This is solved by using the quoted array syntax where it is not used yet.

Resolves: #87970
Releases: master, 9.5
Change-Id: I36f4ee86765fae24f7daa0e7e59ff9fc6741144f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62500
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Patrick Broens <patrick@patrickbroens.nl>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Susanne Moog <look@susi.dev>