[BUGFIX] Set controller object name for widget request
While introducing the fully qualified controller class names
in the extbase plugin configuration the originally used setter
\TYPO3\CMS\Extbase\Mvc\Request::setControllerObjectName() has
no longer been used to guess extension name, subpackage key and
controller name from the class name since all that information
is known. Said setter has been kept nevertheless and it was
overlooked that it was still used by fluid widgets. This leads
to property \TYPO3\CMS\Extbase\Mvc\Request::$controllerObjectName
being empty in widget requests which then leads to an exception
when trying to create a ClassSchema for the controller object
name "".

To fix this, the widget request is now created with the controller
object name as constructor argument.

Releases: master
Resolves: #91418
Change-Id: I6abcdb8c68e831459228cc35c3263cec83d16f67
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64505
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>