[BUGFIX] Do not HTML encode attributes in Script and Asset VH
ScriptViewHelper and AssetViewHelper do not directly output any tag,
but only pass the arguments to AssetCollector.

Therefore the tag attribute values must not be HTML encoded.
To achieve that, we use a subclass of TagBuilder before the tag
is initialised, which does never encode attribute values.

The tag view helper inheritance is kept because the tag builder
nicely abstracts attribute handling, which otherwise would have to
be duplicated.

Also ensure that rendered children will not be encoded as this string
is used as CSS or JavaScript, where HTML encoding won't help or rather
break the code.

Resolves: #92706
Related: #92284
Releases: master, 10.4
Change-Id: I823ad167ccbd52e9e0f3ad9bb3d54f554a985fef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71926
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>