[FEATURE] Add content object related arguments to HtmlViewHelper
The HtmlViewHelper is extended for a couple of new arguments,
used to influence the initialization of the content object.

Following arguments are added:

- "data" (array or object)
- "current"
- "currentValueKey"
- "table"

This allows to properly work with dynamic data in a custom parseFunc.

With the provided data record ("data" argument), the individual
entries can now be accessed using TS:field. With the "current"
argument, TS:current can be filled individually and with the
"currentValueKey" argument, a specific value from the data
record can be set as new TS:current.

The "table" argument is added for completeness, so that the
ContentObjectRenderer can be initialized correctly.

Resolves: #92749
Releases: main
Change-Id: Ibecdb7530a33462dd5464524ac7aff31abb7432b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72803
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>