[BUGFIX] Allow slashes as TS keys and escape dots for generated TS
With the introduction of site settings being exposed
to TypoScript with https://review.typo3.org/64128
it has become important to allow more characters
as TypoScript keys. Allowing a slash in addition should
cover cases where paths are exposed as keys.

Additionally the above change reveals, that the
ArrayUtility::flatten method does not properly handle
array keys that has dots in between, as those must
be escaped to produce a correct result.

This change introduces a new flatten method after
several failed attempts to guess what fixes might
not be breaking.

The new method is used in places where no TypoScript
is used, as in those the weird edge case behaviour
of the flatten method isn't expected anyway.

The new method is marked internal until it is decided
how to proceed with the flatten method.

Resolves: #94646
Releases: master, 10.4
Change-Id: I4ebad8a0beece975702d8601a343aa1fdaaa285c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70115
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Helmut Hummel <typo3@helhum.io>