[BUGFIX] Instantiate ImageService in ImageViewHelper constructor
In #94491 the injection of the ImageService in the
ImageViewHelper was replaced by instantiating
with GeneralUtility. This was done, since the
ImageViewHelper is also used in the install tool,
where no DI of this service is possible.

However, since ImageService can be extended by
third-party ViewHelpers, this then broke when
accessing $this->imageService.

This is now fixed by instantiating  and assigning the
ImageService in the constructor to the $imageService
class variable. This way the DI is still not used, but
backwards compatibility is given.

Resolves: #94728
Related: #94491
Releases: master
Change-Id: Id029a01f7f4f6dabdb2b148ac7b9efcca0f7b0fc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70245
Tested-by: David Steeb <david.steeb@b13.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Ruben Steeb <ruben.steeb@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: David Steeb <david.steeb@b13.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Ruben Steeb <ruben.steeb@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>