[TASK] Re-implement costly regex part of cropHTML programmatically
The magic happening for cropHTML in the regex is
really hard to understand. It led to compilation
failures due to the costly regex pattern applied on
possibly long pieces of text.

In order to replace the "maxCroppingLength"
workaround, where the regex is applied multiple
times with less text, a new implementation is now
in place.

The new implementation extracts the regex part for
html entities to be used as a straight pattern. This
circumvents the bottleneck of the "entity|."
alternative match for everything.

The remaining logic is implemented in a programmatic
way, which is easy to follow from top to bottom with
added comments.

Functionality of `crop()` is moved to new core class
`\TYPO3\CMS\Core\Text\TextCropper` and `cropHTML()`
methods is moved to corresponding new core class
`\TYPO3\CMS\Core\Html\HtmlCropper` with a modernized
method signature. Old internal ContentObjectRenderer
methods are kept as wrapper methods to properly prepare
argument preparation.

Additionally new `HtmlCropper` and `TextCropper` are
conditionally used in the well known fluid viewhelper
`\TYPO3\CMS\Fluid\ViewHelpers\Format\CropViewHelper`.

MultiByte position calculation error for `TextCropper`
is fixed along the way which bubble up using the html
crop tests from `ContentObjectRenderer` also for the
`TextCropper`.

Resolves: #97125
Related: #77642
Releases: main
Change-Id: I15b778370df0d0a3d3087a0b199c3e16f7a46add
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73834
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>