[TASK] Allow EXT: syntax in f:uri.resource VH
f:uri.resource is frequently used to load CSS and
JavaScript resources. It is the only ViewHelper that
does not understand the general "EXT:" syntax that
is usually available when extension resource files
are referenced.

This is now added to improve consistency and
documented as best practice in VH documentation.

Change-Id: I504c79eef349b1ea6cc70cc0d7753319e1f356d9
Resolves: #96414
Releases: main, 11.5
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72759
Tested-by: Helmut Hummel <typo3@helhum.io>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Helmut Hummel <typo3@helhum.io>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>