[BUGFIX] Prevent application type exception in ViewHelper
In case the "f:format.html" ViewHelper is called
in the install tool or via CLI, the request might
miss the application type attribute. This led to
an exception on checking the application type.

This is now fixed by checking the request before
calling the static ApplicationType check.

Resolves: #96788
Releases: main, 11.5
Change-Id: I1e48a3556d510d2af82d4f95828d44c182db0663
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73423
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>