[TASK] Let PageRenderer render a PSR-7 response
In order to avoid the internal HtmlResponse class usage
and to prepare for possible APIs to manage Content Security
Policy headers assets rendered by the PageRenderer it's good
for PageRenderer to be able to render full PSR-7 responses.

Also cleanups up ModuleTemplate to make use of this
capability.

Resolves: #97071
Releases: main
Change-Id: If628d33c916367e204695252b76f629397dd995a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73604
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>