[TASK] Add default value for method in FormViewHelper
The default value for argument "method" is effectively
"post", so this is added as default value in registerArgument
so that it will show up in the automatically generated
viewhelper reference.

Since the argument is always added anyway (after lowercasing it),
there is no difference in the rendered output. The TYPO3 default
differs from the default general default for forms, which is GET.

Resolves: #99086
Releases: main
Change-Id: If34ebc1a65527a605e64513c598a5c381dbeebf6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76599
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>