[DOCS] Remove broken link from ViewHelper PHPDoc
The PHPDoc block in the ViewHelper source is used to render
the ViewHelper documentation. Remove link to Changelog with
incorrect formatting in asset.css ViewHelper, which never worked.

As the Changelog is from version 10.4, there is little
benefit and more risk of linking to outdated information in
the future.

Resolves: #100151
Releases: main, 11.5, 10.4
Change-Id: I0b11e7ec90ccfa8864ca4ddaa7ca83004fe266c0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78106
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>