[DOCS] Improve asset ViewHelper description
The class name is removed since it is not customary for ViewHelpers and
can easily be derived from the ViewHelper name.

A description is added. In particular, the TYPO3 AssetCollector is
mentioned.

More details are added after the example (so the developer can still
easily see the example in a preview of the IDE).

Also, the specific attributes for this element are mentioned because
the automatically generated ViewHelper reference lists all attributes
in alphabetical order and quite a few make no sense for the asset
ViewHelpers.

Resolves: #100285
Releases: main
Change-Id: I17d5f72ca4de4ad4ce4f244de29a54e97eecb7ff
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78242
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>