[BUGFIX] Allow untrusted query arguments in ViewHelpers
With #98488, the typolink option "addQueryString" changed
the behaviour to only add mapped query arguments by default.
To add all - even "untrusted" - arguments to the generated
URL, the new "untrusted" keyword has to be used.

This now adds the possibility to set the new keyword in
all related ViewHelpers. Therefore, the "addQueryString"
property of the extbase UriBuilder is adjusted as well.

Resolves: #100583
Related: #98488
Releases: main
Change-Id: I3a4b62ec734b13e6f3cd721f4a6acefe9877b72b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78804
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>