[DOCS] Clarify usage of "value" with "property" in <f:form.*> ViewHelpers
It is already possible to use the "value" argument together
with "property" in form ViewHelpers. This should be reflected
in the ViewHelper documentation.

One possible use case is a form to edit a user. For the password
field you would want to define property="password", but also
value="" to not expose the current password hash to the form.

Resolves: #102206
Releases: main, 12.4, 11.5
Change-Id: I3c7b1a201e38ccef84efcc863075043cf461dd20
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81459
Tested-by: Simon Praetorius <simon@praetorius.me>
Reviewed-by: Simon Praetorius <simon@praetorius.me>