[TASK] Remove dead catch in ImageViewHelper
RuntimeException is never thrown in the executed code of
ImageViewHelper. Reason is, that the image either is
inside the fallback storage "0" or it was already
caught by InvalidArgumentException catch block.

For reference: There is no way to find out, which
RuntimeException the author meant to catch. I assume it
was the one in AbstractFile->getStorage (code: 1381570091).

Resolves: #102223
Related: #71686
Releases: main, 12.4
Change-Id: Ia08c085a13c29b4188f955887dcaa0410c2f47b0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81502
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>